feat(manifest): Removing 'local' as a term, aligning with workspace
