+++ module/base/src/console.ts
@@ -78,7 +78,7 @@ class $ConsoleManager {
    */
   debug(value: false | string): void {
     if (value !== false) {
-      const active = RuntimeIndex.getModuleList('local', value || '@');
+      const active = RuntimeIndex.getModuleList('workspace', value || '@');
       active.add('@npm:debug');
       this.filter('debug', ctx => active.has(ctx.module));
     } else {
