+++ module/manifest/__index__.ts
@@ -8,4 +8,7 @@ export * from './src/runtime';
 export * from './src/package';
 export * from './src/util';
 export * from './src/file';
-export * from './src/types';
+export * from './src/types/context';
+export * from './src/types/package';
+export * from './src/types/manifest';
+export * from './src/types/common';
