+++ module/manifest/src/delta.ts
@@ -1,12 +1,11 @@
 import fs from 'node:fs/promises';
 
-import {
-  ManifestContext, ManifestModule, ManifestModuleCore, ManifestModuleFile,
-  ManifestModuleFileType, ManifestModuleFolderType, ManifestRoot
-} from './types';
-
 import { ManifestModuleUtil } from './module';
 import { path } from './path';
+
+import type { ManifestModule, ManifestModuleCore, ManifestModuleFile, ManifestRoot } from './types/manifest';
+import type { ManifestModuleFileType, ManifestModuleFolderType } from './types/common';
+import type { ManifestContext } from './types/context';
 
 type DeltaEventType = 'added' | 'changed' | 'removed' | 'missing' | 'dirty';
 type DeltaModule = ManifestModuleCore & { files: Record<string, ManifestModuleFile> };
