+++ module/manifest/src/dependencies.ts
@@ -1,22 +1,14 @@
 import { PackageUtil } from './package';
 import { path } from './path';
-import { ManifestContext, ManifestModuleRole, PackageVisitor, PackageVisitReq, Package, ManifestDepCore } from './types';
 
-export type ModuleDep = ManifestDepCore & {
-  pkg: Package;
-  mainLike?: boolean;
-  sourcePath: string;
-  childSet: Set<string>;
-  parentSet: Set<string>;
-  roleSet: Set<ManifestModuleRole>;
-  /** Dependency is direct to the main module or its part of the workspace global set */
-  topLevel?: boolean;
-};
+import type { PackageVisitor, PackageNode } from './types/package';
+import type { ManifestContext } from './types/context';
+import type { ManifestPackageNode } from './types/manifest';
 
 /**
  * Used for walking dependencies for collecting modules for the manifest
  */
-export class ModuleDependencyVisitor implements PackageVisitor<ModuleDep> {
+export class ManifestPackageVisitor implements PackageVisitor<ManifestPackageNode> {
 
   constructor(public ctx: ManifestContext) {
     this.#mainSourcePath = path.resolve(this.ctx.workspace.path, this.ctx.main.folder);
@@ -35,8 +27,8 @@ export class ModuleDependencyVisitor implements PackageVisitor<ModuleDep> {
   /**
    * Initialize visitor, and provide global dependencies
    */
-  async init(req: PackageVisitReq<ModuleDep>): Promise<PackageVisitReq<ModuleDep>[]> {
-    const pkg = PackageUtil.readPackage(req.sourcePath);
+  async init(node: PackageNode<ManifestPackageNode>): Promise<PackageNode<ManifestPackageNode>[]> {
+    const pkg = PackageUtil.readPackage(node.sourcePath);
     const workspacePkg = PackageUtil.readPackage(this.ctx.workspace.path);
 
     this.#mainLikeModules = new Set([
@@ -45,14 +37,20 @@ export class ModuleDependencyVisitor implements PackageVisitor<ModuleDep> {
     ]);
 
     const globals = Object.keys(workspacePkg.travetto?.build?.withModules ?? [])
-      .map(name => PackageUtil.packageReq<ModuleDep>(PackageUtil.resolvePackagePath(name), name in (workspacePkg.dependencies ?? {}), true));
+      .map(name => PackageUtil.packageReq<ManifestPackageNode>(
+        PackageUtil.resolvePackagePath(name),
+        { prod: name in (workspacePkg.dependencies ?? {}), topLevel: true, }
+      ));
 
     // Capture workspace modules as dependencies
     if (this.ctx.workspace.mono && !this.ctx.main.folder) { // We are at the root of the workspace
       for (const mod of await PackageUtil.resolveWorkspaces(this.ctx)) {
         // Add workspace folders, for tests and docs
         this.#mainLikeModules.add(mod.name);
-        globals.push(PackageUtil.packageReq<ModuleDep>(path.resolve(this.ctx.workspace.path, mod.sourcePath), false, true));
+        globals.push(PackageUtil.packageReq<ManifestPackageNode>(
+          path.resolve(this.ctx.workspace.path, mod.sourcePath),
+          { topLevel: true, workspace: true }
+        ));
       }
     }
 
@@ -62,50 +60,48 @@ export class ModuleDependencyVisitor implements PackageVisitor<ModuleDep> {
   /**
    * Is valid dependency for searching
    */
-  valid(req: PackageVisitReq<ModuleDep>): boolean {
-    return req.sourcePath === this.#mainSourcePath || (
-      (!!req.pkg.travetto || req.pkg.private === true || !req.sourcePath.includes('node_modules'))
-    );
+  valid(node: PackageNode<ManifestPackageNode>): boolean {
+    return node.workspace || !!node.pkg.travetto; // Workspace or travetto module
   }
 
   /**
-   * Create dependency from request
+   * Create manifest package from node
    */
-  create(req: PackageVisitReq<ModuleDep>): ModuleDep {
-    const { pkg, sourcePath } = req;
+  create(node: PackageNode<ManifestPackageNode>): ManifestPackageNode {
+    const { prod, pkg, sourcePath } = node;
     const { name, version } = pkg;
     const main = name === this.ctx.main.name;
     const mainLike = main || this.#mainLikeModules.has(name);
     const internal = pkg.private === true;
-    const local = internal || mainLike || !sourcePath.includes('node_modules');
+    const workspace = node.workspace || mainLike;
+    const topLevel = node.topLevel || mainLike;
 
-    const dep: ModuleDep = {
-      name, version, sourcePath, main, mainLike, local, internal, pkg: req.pkg,
-      parentSet: new Set([]), childSet: new Set([]), roleSet: new Set(req.pkg.travetto?.roles), prod: req.prod,
-      topLevel: req.topLevel || this.#mainLikeModules.has(name)
+    const out: ManifestPackageNode = {
+      name, version, sourcePath, main, mainLike, workspace, internal, pkg, prod, topLevel,
+      parent: new Set([]), children: new Set([]), roles: new Set(pkg.travetto?.roles),
     };
 
-    return dep;
+    return out;
   }
 
   /**
    * Visit dependency
    */
-  visit(req: PackageVisitReq<ModuleDep>, dep: ModuleDep): void {
-    const { parent } = req;
+  visit(node: PackageNode<ManifestPackageNode>, dep: ManifestPackageNode): void {
+    const { parent } = node;
     if (parent && dep.name !== this.ctx.main.name) {
-      dep.parentSet.add(parent.name);
-      parent.childSet.add(dep.name);
+      dep.parent!.add(parent.name);
+      parent.children.add(dep.name);
     }
   }
 
   /**
    * Propagate prod, role information through graph
    */
-  complete(deps: Set<ModuleDep>): Set<ModuleDep> {
-    const mapping = new Map<string, { parent: Set<string>, child: Set<string>, el: ModuleDep }>();
+  complete(deps: Set<ManifestPackageNode>): Set<ManifestPackageNode> {
+    const mapping = new Map<string, { parent: Set<string>, child: Set<string>, el: ManifestPackageNode }>();
     for (const el of deps) {
-      mapping.set(el.name, { parent: new Set(el.parentSet), child: new Set(el.childSet), el });
+      mapping.set(el.name, { parent: new Set(el.parent), child: new Set(el.children), el });
     }
 
     const main = mapping.get(this.ctx.main.name)!;
@@ -114,19 +110,19 @@ export class ModuleDependencyVisitor implements PackageVisitor<ModuleDep> {
     for (const { el } of mapping.values()) {
       if (el === main.el) { continue; }
       if (el.topLevel || main.child.has(el.name)) {
-        if (!el.roleSet.size) {
-          el.roleSet.add('std');
+        if (!el.roles.size) {
+          el.roles.add('std');
         }
         main.child.add(el.name); // Ensure top level is a child of main for propagation
         mapping.get(el.name)!.parent.add(main.el.name);
       } else if (!main.child.has(el.name)) { // Not a direct descendent or top-level
         el.prod = false;
-        el.roleSet.clear(); // Only allow roles via propagation
+        el.roles.clear(); // Only allow roles via propagation
       }
     }
 
     // Color parent
-    main.el.roleSet.add('std');
+    main.el.roles.add('std');
     main.el.prod = true;
 
     while (mapping.size > 0) {
@@ -140,8 +136,8 @@ export class ModuleDependencyVisitor implements PackageVisitor<ModuleDep> {
           const { el: cDep, parent } = mapping.get(c)!;
           parent.delete(el.name); // Remove from child
           if (cDep.name in (el.pkg.dependencies ?? {})) { // If an owned prod-dependency
-            for (const role of el.roleSet) {
-              cDep.roleSet.add(role); // Transfer roles
+            for (const role of el.roles) {
+              cDep.roles.add(role); // Transfer roles
             }
             cDep.prod ||= el.prod; // Allow prod to trickle down as needed
           }
