--- module/manifest/src/types.ts
@@ -1,142 +1,0 @@
-export type NodeModuleType = 'module' | 'commonjs';
-export type NodePackageManager = 'yarn' | 'npm';
-
-export type ManifestModuleFileType = 'typings' | 'ts' | 'js' | 'json' | 'package-json' | 'unknown' | 'fixture' | 'md';
-export type ManifestModuleFolderType =
-  '$root' | '$index' | '$package' |
-  'src' | 'bin' | 'support' | 'resources' | 'test' | 'doc' |
-  'test/fixtures' | 'support/fixtures' | 'support/resources' |
-  '$other' | '$transformer';
-
-export type ManifestModuleRole = 'std' | 'test' | 'doc' | 'compile' | 'build';
-
-export type ManifestModuleFile = [string, ManifestModuleFileType, number] | [string, ManifestModuleFileType, number, ManifestModuleRole];
-export type ManifestDepCore = {
-  name: string;
-  main?: boolean;
-  local?: boolean;
-  version: string;
-  prod: boolean;
-  internal?: boolean;
-};
-export type ManifestModuleCore = ManifestDepCore & {
-  sourceFolder: string;
-  outputFolder: string;
-  roles: ManifestModuleRole[];
-  parents: string[];
-};
-
-export type ManifestModule = ManifestModuleCore & {
-  files: Partial<Record<ManifestModuleFolderType, ManifestModuleFile[]>>;
-};
-
-export type ManifestContext = {
-  workspace: {
-    /** Workspace path for module */
-    path: string;
-    /** The module name for the workspace root */
-    name: string;
-    /** Is the workspace a monorepo? */
-    mono?: boolean;
-    /** The module type of the workspace */
-    type: NodeModuleType;
-    /** The package manager of the workspace */
-    manager: NodePackageManager;
-  };
-  build: {
-    /** Compiler folder, relative to workspace */
-    compilerFolder: string;
-    /** Compiler module folder */
-    compilerModuleFolder: string;
-    /** URL for the compiler server */
-    compilerUrl: string;
-    /** Code output folder, relative to workspace */
-    outputFolder: string;
-    /** Location of development-time tool output */
-    toolFolder: string;
-  };
-  main: {
-    /** Main module for manifest */
-    name: string;
-    /** Folder, relative to workspace for main module */
-    folder: string;
-    /** Description of the main module */
-    description?: string;
-    /** Version of the main module */
-    version: string;
-  };
-};
-
-export type ManifestRoot = ManifestContext & {
-  generated: number;
-  modules: Record<string, ManifestModule>;
-};
-
-export type Package = {
-  name: string;
-  type?: NodeModuleType;
-  version: string;
-  description?: string;
-  license?: string;
-  repository?: {
-    url: string;
-    directory?: string;
-  };
-  author?: {
-    email?: string;
-    name?: string;
-  };
-  main: string;
-  homepage?: string;
-  files?: string[];
-  bin?: Record<string, string>;
-  scripts?: Record<string, string>;
-  engines?: Record<string, string>;
-  keywords?: string[];
-
-  dependencies?: Record<string, string>;
-  devDependencies?: Record<string, string>;
-  peerDependencies?: Record<string, string>;
-  peerDependenciesMeta?: Record<string, { optional?: boolean }>;
-  optionalDependencies?: Record<string, string>;
-  travetto?: {
-    displayName?: string;
-    roles?: ManifestModuleRole[];
-    doc?: {
-      output?: string[];
-      root?: string;
-      baseUrl?: string;
-      outputs?: string[];
-    };
-    build?: Partial<ManifestContext['build']> & {
-      isolated?: boolean;
-      withModules?: Record<string, 'main' | true>;
-    };
-  };
-  workspaces?: string[];
-  private?: boolean;
-  publishConfig?: { access?: 'restricted' | 'public' };
-};
-
-type OrProm<T> = T | Promise<T>;
-
-export type PackageVisitReq<T> = { pkg: Package, prod: boolean, sourcePath: string, parent?: T, topLevel?: boolean };
-export type PackageVisitor<T> = {
-  rootPath: string;
-  init?(req: PackageVisitReq<T>): OrProm<undefined | void | PackageVisitReq<T>[]>;
-  valid?(req: PackageVisitReq<T>): boolean;
-  create(req: PackageVisitReq<T>): OrProm<T>;
-  visit?(req: PackageVisitReq<T>, item: T): OrProm<void>;
-  complete?(values: Set<T>): OrProm<Set<T> | undefined>;
-};
-
-export type PackageWorkspaceEntry = { name: string, sourcePath: string };
-
-export type FunctionMetadata = {
-  id: string;
-  source: string;
-  hash?: number;
-  methods?: Record<string, { hash: number }>;
-  synthetic?: boolean;
-  abstract?: boolean;
-};
