+++ module/manifest/src/types/context.ts
@@ -1,0 +1,40 @@
+import type { NodeModuleType, NodePackageManager } from './common';
+
+export type ManifestContext = {
+  workspace: {
+    /** Workspace path for module */
+    path: string;
+    /** The module name for the workspace root */
+    name: string;
+    /** Is the workspace a monorepo? */
+    mono?: boolean;
+    /** The module type of the workspace */
+    type: NodeModuleType;
+    /** The package manager of the workspace */
+    manager: NodePackageManager;
+    /** The default env name */
+    defaultEnv: string;
+  };
+  build: {
+    /** Compiler folder, relative to workspace */
+    compilerFolder: string;
+    /** Compiler module folder */
+    compilerModuleFolder: string;
+    /** URL for the compiler server */
+    compilerUrl: string;
+    /** Code output folder, relative to workspace */
+    outputFolder: string;
+    /** Location of development-time tool output */
+    toolFolder: string;
+  };
+  main: {
+    /** Main module for manifest */
+    name: string;
+    /** Folder, relative to workspace for main module */
+    folder: string;
+    /** Description of the main module */
+    description?: string;
+    /** Version of the main module */
+    version: string;
+  };
+};
